Changehandler stuff removed. Now editing does not work properly, moving of rooms is not supported by the view component of the editor. Will be fixed soon. Start of implementation of MVC design pattern.

+++ zet/src/util/random/distributions/ExponentialDistribution.java
@@ -49,7 +49,7 @@ public class ExponentialDistribution extends Distribution {
    * @throws IllegalArgumentException if min is smaller than max
    */
   public ExponentialDistribution( double lambda ) throws IllegalArgumentException {
-		super( 0, 5 );
+		super( (double)0, (double)5 );
     setLambda( lambda );
   }
